restructred top nav and layers; fixed NSO codes in admin boundaries; updated nso json files
made a bunch of changes to the terminology used in the navigation and added a few more categories based on goverment feedack.

also fixed a few mismatched NSO codes in the boundary file. also added correct mongolian text to the same file

updated multiple nso json files after finding a bug in the API request